+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - ModuleNotFoundError: No module named 'pyairmore.services' | [#9](https://github.com/erayerdin/pyairmore/issues/9)
  - Cache Travis Dependencies | [#11](https://github.com/erayerdin/pyairmore/issues/11)
  - Adding Requests Upgrade to Pip in Travis | [#10](https://github.com/erayerdin/pyairmore/issues/10)
+ - Importing `services.device` Package Fails in PIL is not Found | [#8](https://github.com/erayerdin/pyairmore/issues/8)
 
 ## [0.1.0a12] - 2018-10-20
 ### Changed