+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1a2] - 2018-11-04
+### Removed
+ - `scripts` from `setup.py`
+
 ## [0.1.1a1] - 2018-11-04
 ### Added
  - Test server for testing